Migration vers la SDL 2.0

+++ jeu/box.h
@@ -11,12 +11,12 @@ class box : public Objet3DDeformable{
 private:
 
 	// Hauteur et largeur de la boite
-	int largeur;
-	int hauteur;
+	float largeur;
+	float hauteur;
 
 
 public:
-	box(float16 positionX = 0.0f, float16 positionY = 0.0f, float16 positionZ = 0.0f, float16 angleHorizontal = 0.0f, float16 angleVertical = 0.0f, btScalar mass = 0.0f, int plargeur = 1, int phauteur = 1);
+	box(float16 positionX = 0.0f, float16 positionY = 0.0f, float16 positionZ = 0.0f, float16 angleHorizontal = 0.0f, float16 angleVertical = 0.0f, btScalar mass = 0.0f, float plargeur = 1, float phauteur = 1);
 	virtual ~box();
 
 	void dessinerBox(float x, float y, float z);
